Inform of auto-generated columns in migrations
When running migrations, I attempted to add the `id` column as well as the `created_at` and `updated_at` columns and ran into errors because of the migrations already creating those columns for me. I do realize that this is probably a Fizz feature, but it would be good to know.