Make the buffalo verification use the "help" subcommand
As of version v0.16.17, the `buffalo` command with no subcommand returns an error after printing out the help message
e.g. `ERRO[0000] Error: subcommand is required`

Thus adding the help subcommand avoids displaying that error to the new
buffalo user.

I also updated the output of the help command to the match the latest version.