Add notes to make it obvious how to get Sqlite support
After spending over an hour trying to figure out why `sqlite` was not available as
a --db-type with the `buffalo` command I finally realized that the section that said "Custom installation with SQLite3 Support" meant "to get Sqlite support you MUST do a custom installation" instead of what I initially thought it meant which was "when doing a custom installation this is the way to do it with Sqlite support vs. the next section which shows how to do it without Sqlite support." This pull request is my attempt to help the next person who comes along and is as confused by the docs as I was.