Fix `go get` commands in instructions
The "v" is trimmed from the git tag which is correct for the prebuilt binary file names (for some reason) but not the correct Git tag.  This means the `go get` commands are broken but the other commands on the install page work.

I think the binary archives/file should include the "v" in the package name and the docs package should stop trimming it.